+++ tests/testthat/test-handle_ru_geotraces.R
@@ -122,4 +122,29 @@ test_that("handle_ru_geotraces annotates WKT lines with lon lat alt (no acc)", {
   }
 })
 
+test_that("handle_ru_geotraces removes last empty coordinate from GJ", {
+  data("geo_fs") # parse T
+  data("geo_gj") # parse T, wkt T
+
+
+  geo_fields <- geo_fs %>%
+    # dplyr::filter(type %in% c("geopoint", "geotrace", "geoshape")) %>%
+    dplyr::filter(type == "geotrace") %>%
+    magrittr::extract2("ruodk_name")
+
+  for (i in seq_len(length(geo_fields))) {
+    len_coords <- length(geo_gj[[geo_fields[i]]][[1]]$coordinates)
+
+    # Cmpty coord is list(NULL, NULL) which compacts to list() of length 0
+    last_coord_compact <- purrr::compact(
+      geo_gj[[geo_fields[i]]][[1]]$coordinates[[len_coords]]
+    )
+    # WKT geofields must not contain bad_coord
+    testthat::expect_false(
+      length(last_coord_compact) == 0,
+      label = glue::glue("GJ field {geo_fields[i]} last coord is not empty")
+    )
+  }
+})
+
 # usethis::use_r("handle_ru_geotraces") # nolint